+++ pkg/proxy/integrations/mysqlparser/comFetchPacket.go
