+++ pkg/proxy/integrations/mysqlparser/comStmtPrepareOk.go
@@ -1,0 +1,190 @@
+package mysqlparser
+
+import (
+	"bytes"
+	"encoding/binary"
+	"errors"
+
+	"go.keploy.io/server/pkg/models"
+)
+
+type StmtPrepareOk struct {
+	Status       byte               `yaml:"status"`
+	StatementID  uint32             `yaml:"statement_id"`
+	NumColumns   uint16             `yaml:"num_columns"`
+	NumParams    uint16             `yaml:"num_params"`
+	WarningCount uint16             `yaml:"warning_count"`
+	ColumnDefs   []ColumnDefinition `yaml:"column_definitions"`
+	ParamDefs    []ColumnDefinition `yaml:"param_definitions"`
+}
+
+func decodeComStmtPrepareOk(data []byte) (*StmtPrepareOk, error) {
+	if len(data) < 12 {
+		return nil, errors.New("data length is not enough for COM_STMT_PREPARE_OK")
+	}
+
+	response := &StmtPrepareOk{
+		Status:       data[0],
+		StatementID:  binary.LittleEndian.Uint32(data[1:5]),
+		NumColumns:   binary.LittleEndian.Uint16(data[5:7]),
+		NumParams:    binary.LittleEndian.Uint16(data[7:9]),
+		WarningCount: binary.LittleEndian.Uint16(data[10:12]),
+	}
+
+	offset := 12
+
+	if response.NumParams > 0 {
+		for i := uint16(0); i < response.NumParams; i++ {
+			columnDef := ColumnDefinition{}
+			columnHeader := PacketHeader{
+				PacketLength:     data[offset],
+				PacketSequenceID: data[offset+3],
+			}
+			columnDef.PacketHeader = columnHeader
+			offset += 4 //Header of packet
+			var err error
+			columnDef.Catalog, err = readLengthEncodedString(data, &offset)
+			if err != nil {
+				return nil, err
+			}
+			columnDef.Schema, err = readLengthEncodedString(data, &offset)
+			if err != nil {
+				return nil, err
+			}
+			columnDef.Table, err = readLengthEncodedString(data, &offset)
+			if err != nil {
+				return nil, err
+			}
+			columnDef.OrgTable, err = readLengthEncodedString(data, &offset)
+			if err != nil {
+				return nil, err
+			}
+			columnDef.Name, err = readLengthEncodedString(data, &offset)
+			if err != nil {
+				return nil, err
+			}
+			columnDef.OrgName, err = readLengthEncodedString(data, &offset)
+			if err != nil {
+				return nil, err
+			}
+			offset++ //filler
+			columnDef.CharacterSet = binary.LittleEndian.Uint16(data[offset : offset+2])
+			columnDef.ColumnLength = binary.LittleEndian.Uint32(data[offset+2 : offset+6])
+			columnDef.ColumnType = data[offset+6]
+			columnDef.Flags = binary.LittleEndian.Uint16(data[offset+7 : offset+9])
+			columnDef.Decimals = data[offset+9]
+			offset += 10
+			offset += 2 // filler
+			response.ParamDefs = append(response.ParamDefs, columnDef)
+		}
+		offset += 9 //skip EOF packet for Parameter Definition
+	}
+
+	if response.NumColumns > 0 {
+		for i := uint16(0); i < response.NumColumns; i++ {
+			columnDef := ColumnDefinition{}
+			columnHeader := PacketHeader{
+				PacketLength:     data[offset],
+				PacketSequenceID: data[offset+3],
+			}
+			columnDef.PacketHeader = columnHeader
+			offset += 4
+			var err error
+			columnDef.Catalog, err = readLengthEncodedString(data, &offset)
+			if err != nil {
+				return nil, err
+			}
+			columnDef.Schema, err = readLengthEncodedString(data, &offset)
+			if err != nil {
+				return nil, err
+			}
+			columnDef.Table, err = readLengthEncodedString(data, &offset)
+			if err != nil {
+				return nil, err
+			}
+			columnDef.OrgTable, err = readLengthEncodedString(data, &offset)
+			if err != nil {
+				return nil, err
+			}
+			columnDef.Name, err = readLengthEncodedString(data, &offset)
+			if err != nil {
+				return nil, err
+			}
+			columnDef.OrgName, err = readLengthEncodedString(data, &offset)
+			if err != nil {
+				return nil, err
+			}
+			offset++ //filler
+			columnDef.CharacterSet = binary.LittleEndian.Uint16(data[offset : offset+2])
+			columnDef.ColumnLength = binary.LittleEndian.Uint32(data[offset+2 : offset+6])
+			columnDef.ColumnType = data[offset+6]
+			columnDef.Flags = binary.LittleEndian.Uint16(data[offset+7 : offset+9])
+			columnDef.Decimals = data[offset+9]
+			offset += 10
+			offset += 2 // filler
+			response.ColumnDefs = append(response.ColumnDefs, columnDef)
+		}
+		offset += 9 //skip EOF packet for Column Definitions
+	}
+
+	return response, nil
+}
+
+func encodeStmtPrepareOk(packet *models.MySQLStmtPrepareOk) ([]byte, error) {
+	buf := &bytes.Buffer{}
+	buf.Write([]byte{0x0C, 0x00, 0x00, 0x01})
+	// Encode the Status field
+	if err := binary.Write(buf, binary.LittleEndian, uint8(packet.Status)); err != nil {
+		return nil, err
+	}
+
+	// Encode the StatementID field
+	if err := binary.Write(buf, binary.LittleEndian, packet.StatementID); err != nil {
+		return nil, err
+	}
+
+	// Encode the NumColumns field
+	if err := binary.Write(buf, binary.LittleEndian, uint16(packet.NumColumns)); err != nil {
+		return nil, err
+	}
+
+	// Encode the NumParams field
+	if err := binary.Write(buf, binary.LittleEndian, uint16(packet.NumParams)); err != nil {
+		return nil, err
+	}
+
+	// Encode the WarningCount field
+	if err := binary.Write(buf, binary.LittleEndian, uint16(packet.WarningCount)); err != nil {
+		return nil, err
+	}
+
+	buf.WriteByte(0x00) // Reserved byte
+
+	seqNum := byte(2)
+	for i := uint16(0); i < packet.NumParams; i++ {
+		param := packet.ParamDefs[i]
+		if err := encodeColumnDefinition(buf, &param, &seqNum); err != nil {
+			return nil, err
+		}
+	}
+	if packet.NumParams > 0 {
+		// Write EOF marker for parameter definitions
+		buf.Write([]byte{5, 0, 0, seqNum, 0xFE, 0x00, 0x00, 0x02, 0x00})
+		seqNum++
+	}
+
+	// Encode column definitions
+	for _, col := range packet.ColumnDefs {
+		if err := encodeColumnDefinition(buf, &col, &seqNum); err != nil {
+			return nil, err
+		}
+	}
+
+	if packet.NumColumns > 0 {
+		// Write EOF marker for column definitions
+		buf.Write([]byte{5, 0, 0, seqNum, 0xFE, 0x00, 0x00, 0x02, 0x00})
+		seqNum++
+	}
+
+	return buf.Bytes(), nil
+}
