+++ pkg/proxy/integrations/mysqlparser/comStmtSendLongDataPacket.go
@@ -1,0 +1,26 @@
+package mysqlparser
+
+import (
+	"encoding/binary"
+	"fmt"
+)
+
+type COM_STMT_SEND_LONG_DATA struct {
+	StatementID uint32 `yaml:"statement_id"`
+	ParameterID uint16 `yaml:"parameter_id"`
+	Data        []byte `yaml:"data"`
+}
+
+func decodeComStmtSendLongData(packet []byte) (COM_STMT_SEND_LONG_DATA, error) {
+	if len(packet) < 7 || packet[0] != 0x18 {
+		return COM_STMT_SEND_LONG_DATA{}, fmt.Errorf("invalid COM_STMT_SEND_LONG_DATA packet")
+	}
+	stmtID := binary.LittleEndian.Uint32(packet[1:5])
+	paramID := binary.LittleEndian.Uint16(packet[5:7])
+	data := packet[7:]
+	return COM_STMT_SEND_LONG_DATA{
+		StatementID: stmtID,
+		ParameterID: paramID,
+		Data:        data,
+	}, nil
+}
